Extract from_xml from process
